For several distribution functions it was unclear what the parameters mean, as many distributions have two forms: a parameter either indicates the rate, or the average (1/rate). Most distributions have been described in much more detail, and the unit tests for these distribution have been switched on. The ones clarified in this version are: NegBinomial, DistPearson5, DistPearson6, Poisson, Triangular, Uniform, and Weibull.

+++ dsol-core/src/main/java/nl/tudelft/simulation/jstats/distributions/DistBinomial.java
@@ -4,8 +4,10 @@ import nl.tudelft.simulation.jstats.math.ProbMath;
 import nl.tudelft.simulation.jstats.streams.StreamInterface;
 
 /**
- * The Binomial distribution. For more information on this distribution see
- * <a href="https://mathworld.wolfram.com/BinomialDistribution.html"> https://mathworld.wolfram.com/BinomialDistribution.html </a>
+ * The Binomial distribution. The binomial distribution is the probability of the number of successes in a sequence of n
+ * independent experiments, each with success (probability p) or failure (probability q = 1 − p). For more information on this
+ * distribution see <a href="https://mathworld.wolfram.com/BinomialDistribution.html">
+ * https://mathworld.wolfram.com/BinomialDistribution.html </a>
  * <p>
  * Copyright (c) 2002-2021 Delft University of Technology, Jaffalaan 5, 2628 BX Delft, the Netherlands. All rights reserved. See
  * for project information <a href="https://simulation.tudelft.nl/" target="_blank"> https://simulation.tudelft.nl</a>. The DSOL
@@ -21,20 +23,21 @@ public class DistBinomial extends DistDiscrete
     /** */
     private static final long serialVersionUID = 1L;
 
-    /** n is the n-parameter of the Binomial distribution. */
-    private final long n;
+    /** n is the number of independent experiments for the Binomial distribution. */
+    private final int n;
 
-    /** p is the p-value of the binomial distribution. */
+    /** p is the probability of success for each individual trial in the binomial distribution. */
     private final double p;
 
     /**
-     * constructs a Binomial distribution. Number of successes in n independent Bernoulli trials with probability p of success
-     * on each trial.
+     * constructs a Binomial distribution. It calculates the probability for a number of successes in n independent Bernoulli
+     * trials with probability p of success on each trial.
      * @param stream StreamInterface; the random number stream
-     * @param n long; is the n-parameter of the Binomial distribution
-     * @param p double; is the p-parameter of the Binomial distribution
+     * @param n long; the number of independent experiments for the Binomial distribution
+     * @param p double; the probability of success for each individual trial in the binomial distribution
+     * @throws IllegalArgumentException when n &lt;= 0 or p &lt;= 0 or p &gt;= 1
      */
-    public DistBinomial(final StreamInterface stream, final long n, final double p)
+    public DistBinomial(final StreamInterface stream, final int n, final double p)
     {
         super(stream);
         if ((n > 0) && (p > 0) && (p < 1))
@@ -53,7 +56,7 @@ public class DistBinomial extends DistDiscrete
     public long draw()
     {
         long x = 0;
-        for (long i = 0; i < this.n; i++)
+        for (int i = 0; i < this.n; i++)
         {
             if (this.stream.nextDouble() <= this.p)
             {
@@ -69,22 +72,24 @@ public class DistBinomial extends DistDiscrete
     {
         if (observation <= this.n && observation >= 0)
         {
-            return ProbMath.combinations((int) this.n, observation) * Math.pow(this.p, observation)
+            return ProbMath.combinations(this.n, observation) * Math.pow(this.p, observation)
                     * Math.pow(1 - this.p, this.n - observation);
         }
         return 0.0;
     }
 
     /**
-     * @return n
+     * Return the number of independent experiments for the Binomial distribution.
+     * @return int; the number of independent experiments for the Binomial distribution
      */
-    public final long getN()
+    public final int getN()
     {
         return this.n;
     }
 
     /**
-     * @return p
+     * Return the probability of success for each individual trial in the binomial distribution.
+     * @return double; the probability of success for each individual trial in the binomial distribution
      */
     public final double getP()
     {
